Ability to segment sub-documents.
Right now n-grams are generated over the entire document. However,
that may not be a realistic way in which it happens over sound. There
could be several breaks within a document; the analogy here being not
all continuous windows should be reported.

The GateKeeper interface allows for various strategies of "leave out"
to be implemented.