Infrastructure for ngram distortion.
To be more like audio output, it is possible that characters should be
"distorted" in some way before being added to the suffix tree. An
example of this already exists via the GateKeeper interface; this
package will eventually supercede that.

During distortion, the package also identifies the legality of an
ngram; see, for instance, the Sentence transformer versus the rest. In
this way, the package serves a dual purpose.

The package defines a decorator infrastructure -- new classes should
heed the layout.